requirements.txt
Flask==1.1.2
Flask-API==2.0
viaa-chassis==0.1.3
uWSGI==2.0.18